(#47) Preliminary support for native packages.

+++ lib/freight/apt.sh
@@ -107,10 +107,10 @@ apt_cache() {
 			*.deb) apt_cache_binary "$DIST" "$DISTCACHE" "$PATHNAME" "$COMP" "$PACKAGE";;
 
 			# Source packages.  The *.dsc file is considered the "entrypoint"
-			# and will find the associated *.orig.tar.gz and *.diff.gz as
-			# they are needed.
+			# and will find the associated *.orig.tar.gz, *.diff.gz, and/or
+            # *.tar.gz as they are needed.
 			*.dsc) apt_cache_source "$DIST" "$DISTCACHE" "$PATHNAME" "$COMP" "$PACKAGE";;
-			*.debian.tar.gz|*.diff.gz|*.orig.tar.gz|*.deb-control|*.dsc-cached) ;;
+			*.debian.tar.gz|*.diff.gz|*.orig.tar.gz|*.tar.gz|*.deb-control|*.dsc-cached) ;;
 
 			*) echo "# [freight] skipping extraneous file $PATHNAME" >&2;;
 		esac
@@ -319,9 +319,9 @@ EOF
 
 }
 
-# Add a source package to the given dist and to the pool.  *.orig.tar.gz and
-# *.diff.gz will be found based on PATHNAME and associated with the correct
-# source package.
+# Add a source package to the given dist and to the pool.  *.orig.tar.gz,
+# *.diff.gz, and/or *.tar.gz will be found based on PATHNAME and associated
+# with the correct source package.
 apt_cache_source() {
 	DIST="$1"
 	DISTCACHE="$2"
@@ -337,15 +337,16 @@ apt_cache_source() {
 	DEBTAR_FILENAME="${NAME}_${VERSION%*:}.debian.tar.gz"
 	DIFFGZ_FILENAME="${NAME}_${VERSION%*:}.diff.gz"
 	ORIG_FILENAME="${NAME}_${ORIG_VERSION}.orig.tar.gz"
+	TAR_FILENAME="${NAME}_${VERSION%*:}.tar.gz"
 
+    # Find which style of diff they're using.
 	if [ -f "$VARLIB/apt/$DIST/$DIRNAME/$DEBTAR_FILENAME" ]
 	then DIFF_FILENAME=${DEBTAR_FILENAME}
 	else DIFF_FILENAME=${DIFFGZ_FILENAME}
 	fi
 
 	# Verify this package by ensuring the other necessary files are present.
-	[ -f "$VARLIB/apt/$DIST/$DIRNAME/$ORIG_FILENAME" \
-	-a -f "$VARLIB/apt/$DIST/$DIRNAME/$DIFF_FILENAME" ] || {
+	[ -f "$VARLIB/apt/$DIST/$DIRNAME/$ORIG_FILENAME" -a -f "$VARLIB/apt/$DIST/$DIRNAME/$DIFF_FILENAME" -o -f "$VARLIB/apt/$DIST/$DIRNAME/$TAR_FILENAME" ] || {
 		echo "# [freight] skipping invalid Debian source package $PATHNAME" >&2
 		return
 	}
@@ -358,11 +359,12 @@ apt_cache_source() {
 	# if it isn't already there (which can happen when two packages derive
 	# from the same original tarball).
 	mkdir -p "$DISTCACHE/.refs/$COMP"
-	for FILENAME in "$DSC_FILENAME" "$ORIG_FILENAME" "$DIFF_FILENAME"
+	for FILENAME in "$DSC_FILENAME" "$ORIG_FILENAME" "$DIFF_FILENAME" "$TAR_FILENAME"
 	do
-		[ -f "$DISTCACHE/.refs/$COMP/$FILENAME" ] ||
-		ln "$VARLIB/apt/$DIST/$DIRNAME/$FILENAME" "$DISTCACHE/.refs/$COMP" ||
-		cp "$VARLIB/apt/$DIST/$DIRNAME/$FILENAME" "$DISTCACHE/.refs/$COMP"
+        [ -f "$VARLIB/apt/$DIST/$DIRNAME/$FILENAME" ] || continue
+        [ -f "$DISTCACHE/.refs/$COMP/$FILENAME" ] ||
+        ln "$VARLIB/apt/$DIST/$DIRNAME/$FILENAME" "$DISTCACHE/.refs/$COMP" ||
+        cp "$VARLIB/apt/$DIST/$DIRNAME/$FILENAME" "$DISTCACHE/.refs/$COMP"
 	done
 
 	# Package properties.  Remove the epoch from the version number
@@ -371,9 +373,9 @@ apt_cache_source() {
 	# Link this source package into the pool.
 	POOL="pool/$DIST/$COMP/$(apt_prefix "$NAME")/$NAME"
 	mkdir -p "$VARCACHE/$POOL"
-	for FILENAME in "$DSC_FILENAME" "$ORIG_FILENAME" "$DIFF_FILENAME"
+	for FILENAME in "$DSC_FILENAME" "$ORIG_FILENAME" "$DIFF_FILENAME" "$TAR_FILENAME"
 	do
-		if [ ! -f "$VARCACHE/$POOL/$FILENAME" ]
+		if [ -f "$DISTCACHE/.refs/$COMP/$FILENAME" -a ! -f "$VARCACHE/$POOL/$FILENAME" ]
 		then
 			echo "# [freight] adding $FILENAME to pool" >&2
 			ln "$DISTCACHE/.refs/$COMP/$FILENAME" "$VARCACHE/$POOL"
@@ -395,22 +397,25 @@ apt_cache_source() {
 			sed "s/^Source:/Package:/"
 			echo "Directory: DIRECTORY"
 			echo "Files:"
-			for FILENAME in "$DSC_FILENAME" "$ORIG_FILENAME" "$DIFF_FILENAME"
+			for FILENAME in "$DSC_FILENAME" "$ORIG_FILENAME" "$DIFF_FILENAME" "$TAR_FILENAME"
 			do
+                [ -f "$VARCACHE/$POOL/$FILENAME" ] || continue
 				SIZE="$(apt_filesize "$VARCACHE/$POOL/$FILENAME")"
 				MD5="$(apt_md5 "$VARCACHE/$POOL/$FILENAME")"
 				echo " $MD5 $SIZE $FILENAME"
 			done
 			echo "Checksums-Sha1:"
-			for FILENAME in "$DSC_FILENAME" "$ORIG_FILENAME" "$DIFF_FILENAME"
+			for FILENAME in "$DSC_FILENAME" "$ORIG_FILENAME" "$DIFF_FILENAME" "$TAR_FILENAME"
 			do
+                [ -f "$VARCACHE/$POOL/$FILENAME" ] || continue
 				SIZE="$(apt_filesize "$VARCACHE/$POOL/$FILENAME")"
 				SHA1="$(apt_sha1 "$VARCACHE/$POOL/$FILENAME")"
 				echo " $SHA1 $SIZE $FILENAME"
 			done
 			echo "Checksums-Sha256:"
-			for FILENAME in "$DSC_FILENAME" "$ORIG_FILENAME" "$DIFF_FILENAME"
+			for FILENAME in "$DSC_FILENAME" "$ORIG_FILENAME" "$DIFF_FILENAME" "$TAR_FILENAME"
 			do
+                [ -f "$VARCACHE/$POOL/$FILENAME" ] || continue
 				SIZE="$(apt_filesize "$VARCACHE/$POOL/$FILENAME")"
 				SHA256="$(apt_sha256 "$VARCACHE/$POOL/$FILENAME")"
 				echo " $SHA256 $SIZE $FILENAME"
